replace from controller to utils
